Add infrastructure for building scala-library-jribble.jar.
At the moment, only scala.ScalaObject will be compiled into jribble
and packaged into scala-library-jribble.jar because scalac is not
ready to compile entire library into jribble, yet.

The scala-library-jribble.jar will be put in build/pack directory
as it works with scala-library.jar.

Put ScalaLibrary.gwt.xml in src/library. Not sure if this is the best
location but couldn't find better one without polluting top-level
directory list.

ScalaLibrary.gwt.xml is needed by GWT to discover that given jar is a
module so it will try to scan it's contents and parse discovered
jribble files.

Change-Id: I25c2f03682dae8b32f7bb6e813100f77e5396797
Signed-off-by: Grzegorz Kossakowski <grek@google.com>