Fixed two issues related to printing jribble for object declarations.
First problem was with initializing $MODULE field. The expression which
was meant to be a new call was broken.

Second problem was with emitting constructors for module classes. The
name of the constructor didn't match class name.

These kind of problems make it quite apparent that handling of modules
should be done in earlier phases and they should be translated into
classes. Perhaps code used for jvm backend can be reused.
Need to investigate it.

Added test case covering these problems and corrected hello.scala test.

Change-Id: I4bd7af53f87337bf7ce5dcae30a4865a5176ff76
Signed-off-by: Grzegorz Kossakowski <grek@google.com>