Made it more apparent which cases for trees we support in jribble.
Made pattern matching for new calls, super calls and regular calls
slightly more clear. We rely more on pattern matching on trees than
inspecting symbols which seems to be a good idea.

Got rid of underscore in pattern matching on supported trees in jirbble
printer and replace it with direct matches that still delegate handling
of those cases to the super class. This way it's more apparent what's
handled in direct way and what's delegated.

Checked list of possible tree types we should support and added missing
PackageDef (which is already supported in the code). This way the list
is complete.

Change-Id: Ia8b3644413d9079719f353fc14f781fa8ef224d8
Signed-off-by: Grzegorz Kossakowski <grek@google.com>