Checked implementation of try { .. } catch { .. } handling.
Basic functionality for try { .. } catch { .. } works correctly but
in more sohpisticated case compilation to jribble will crash due to
problems with pattern matching.

Made this crashes more clear so it's easier to understand why compiler
did crash. These crashes are essentially related to a problem that pattern
matching logic can emit trees that are very hard to translate to jribble.
We probably crash in cases were this translation would be probably fairly
easy too, but this restriction can be lifted later on.

Added basic test of try functionality.

Change-Id: Icbedab2cfa0257ba01332fd3d3c5f368693247e3