Implemented handling of non-forward jumps.
Non-forward jumps are being translated into while loops with continue
instruction acting as a jump. Also, LabelDefs can have a list of
parameters which are just local variables. Whenever we make a jump
we need to assign correct values to them before making a jump.

Added a test-case which contains a tail-recursive method definition
that is being transformed into LabelDef and non-forward jump.

Change-Id: I5e591ced6243b6104da733f93e13ef3a69fa9f4d
Signed-off-by: Grzegorz Kossakowski <grek@google.com>