Normalize trees so they don't include private local vals.
Scala allows one to have val declared in method scope with 'private'
modifier. This is not supported in jribble so we get rid of 'private'
modifier on vals that are declared locally.

Added a pattern matching constructs that triggers scala compiler to
create a tree containing private vals so we can verify that we
correctly get rid of private modifier.

Change-Id: Iddd2e73d5692fd6a3b17eaa9d9f36dd6c37adafa
Signed-off-by: Grzegorz Kossakowski <grek@google.com>