Fixed handling construction of multidimensional arrays.
Fixed handling construction of multidimensional arrays by
extracting the type of element held in the array and number of
it's dimensions.

Moreover, we print expression for the size of the array in each
dimension. If there is no such expression we'll print just []
to still denote correct number of dimensions.

Added test-case for situation that number of expressions does
not match number of dimensions of the array.

Change-Id: I4645d28f23d0407cdc28d9686aaa86e3c4d239d6
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>