Use Jribble-specific compiler options for compiling scala library.
In addition to Jribble files we need class files in order to compile
code using gwt. Since Jribble files need to be consistent with class
files we are using the same compiler options for both.

Need for class files is temporary and in the future this change will
be reverted.

Change-Id: Ief78de7df958fcc6a58180edb7a87b29b59550ad
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>