Fix wrong handling of LabelDefs of Nothing type.
LabelDefs of Nothing type would have a synthetic
ValDef introduced to carry it's result. This is
wrong because we always discard assignments to
with rhs of Nothing type.

It's worth noting that the whole logic is still
broken but not in this particular case. The problem
is with LabelDefs changing it type while being
rewritten. It's confusing the rest of the logic
in NormalizeForJribble. I think it's a good
reason to split this logic into two pieces and
perform two tree traversals instead of one.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>