Expanded the range of a warning, and made suppressed warnings visible.
Modified the positioning of "permanently hidden" errors so that when
there is more than one, the later ones are not ignored. Also changed the
error suppression code to emit the error anyway if -Ydebug was given (it
is prefixed with "[suppressed] ".) Since I can't be the only one who
wondered where his errors were vanishing to. No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25485 5e8d7ff9-d8ef-0310-90f0-a4852d11357a