Hand specialized SettingValue.
Discovered every time we do something like

  if (settings.debug.value)

the boolean is coming out of a box. How uncouth. To fix this, I had to
make the storage abstract, so concrete setting types have to declare
the storage personally. This seems a small price to pay. I tried to use
specialization but I think it's impossible to get the type parameter and
the abstract type to agree with one another when mr. invariant var is
the object of your affection, without scalac putting the kibosh on the
whole adventure. No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25755 5e8d7ff9-d8ef-0310-90f0-a4852d11357a